pyproject.toml
[tool.poetry]
name = "git"
version = "0.1.0"
description = ""
authors = ["dunossauro <mendesxeduardo@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.2"
gunicorn = "^20.0.4"
Flask = "^1.1.2"
validate-docbr = "^1.8.2"
flasgger = "^0.9.5"
Pillow = "^8.1.2"
pytesseract = "^0.3.7"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
